Moved settings related services to own file
